move files under spectra_matching/gateways to spectra_matching/storage
delete spectra_matching/gateways directory

--- don't change below this line ---
MLOPS-449 #time 42m
