+++ omigami/ms2deepscore/gateways/fs_data_gateway.py
@@ -12,7 +12,7 @@ from ms2deepscore import SpectrumBinner
 from ms2deepscore.models import SiameseModel
 from tensorflow.python.keras.saving import hdf5_format
 
-from omigami.spectra_matching.gateways.fs_data_gateway import FSDataGateway
+from omigami.spectra_matching.storage import FSDataGateway
 
 
 class MS2DeepScoreFSDataGateway(FSDataGateway):
