+++ omigami/ms2deepscore/gateways/redis_spectrum_gateway.py
@@ -7,9 +7,7 @@ from typing import List
 from ms2deepscore import BinnedSpectrum
 
 from omigami.config import IonModes
-from omigami.spectra_matching.gateways.redis_spectrum_data_gateway import (
-    RedisSpectrumDataGateway,
-)
+from omigami.spectra_matching.storage import RedisSpectrumDataGateway
 from omigami.ms2deepscore.config import BINNED_SPECTRUM_HASHES
 from omigami.ms2deepscore.config import PROJECT_NAME
