+++ omigami/ms2deepscore/tasks/train_model.py
@@ -4,7 +4,7 @@ from typing import List, Dict
 from prefect import Task
 
 from omigami.config import IonModes
-from omigami.spectra_matching.gateways import DataGateway
+from omigami.spectra_matching.storage import DataGateway
 from omigami.ms2deepscore.gateways.redis_spectrum_gateway import (
     MS2DeepScoreRedisSpectrumDataGateway,
 )
