+++ omigami/spec2vec/factory.py
@@ -30,8 +30,7 @@ from omigami.spec2vec.flows.training_flow import (
 from omigami.spec2vec.gateways.redis_spectrum_document import (
     RedisSpectrumDocumentDataGateway,
 )
-from omigami.spectra_matching.gateways import RedisSpectrumDataGateway
-from omigami.spectra_matching.gateways.fs_data_gateway import FSDataGateway
+from omigami.spectra_matching.storage import RedisSpectrumDataGateway, FSDataGateway
 from omigami.spectra_matching.spectrum_cleaner import SpectrumCleaner
