+++ omigami/spec2vec/gateways/fs_document_iterator.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import List
 
-from omigami.spectra_matching.gateways.fs_data_gateway import FSDataGateway
+from omigami.spectra_matching.storage import FSDataGateway
 
 
 class FileSystemDocumentIterator:
