+++ omigami/spec2vec/tasks/process_spectrum/task.py
@@ -3,8 +3,7 @@ from typing import Set, List
 
 from prefect import Task
 
-from omigami.spectra_matching.gateways import RedisSpectrumDataGateway
-from omigami.spectra_matching.gateways.fs_data_gateway import FSDataGateway
+from omigami.spectra_matching.storage import RedisSpectrumDataGateway, FSDataGateway
 from omigami.spec2vec.entities.spectrum_document import SpectrumDocumentData
 from omigami.spec2vec.gateways.spectrum_document import SpectrumDocumentDataGateway
 from omigami.spec2vec.helper_classes.progress_logger import TaskProgressLogger
