+++ omigami/spec2vec/tasks/train_model.py
@@ -10,7 +10,7 @@ from spec2vec.model_building import (
     learning_rates_to_gensim_style,
 )
 
-from omigami.spectra_matching.gateways.fs_data_gateway import FSDataGateway
+from omigami.spectra_matching.storage import FSDataGateway
 from omigami.spec2vec.gateways.fs_document_iterator import FileSystemDocumentIterator
 from omigami.spec2vec.helper_classes.train_logger import (
     CustomTrainingProgressLogger,
