+++ omigami/spectra_matching/storage/__init__.py
@@ -1,3 +1,7 @@
-from omigami.spectra_matching.storage.redis import RedisDataGateway
+from omigami.spectra_matching.storage.data_gateway import DataGateway
+from omigami.spectra_matching.storage.fs_data_gateway import FSDataGateway, KEYS
 from omigami.spectra_matching.storage.redis import REDIS_DB
-from omigami.spectra_matching.storage.redis import REDIS_HOST
+from omigami.spectra_matching.storage.redis import RedisDataGateway
+from omigami.spectra_matching.storage.redis_spectrum_data_gateway import (
+    RedisSpectrumDataGateway,
+)
