+++ omigami/test/conftest.py
@@ -37,7 +37,7 @@ from omigami.spec2vec.entities.embedding import Spec2VecEmbedding
 from omigami.spec2vec.gateways.redis_spectrum_document import (
     RedisSpectrumDocumentDataGateway,
 )
-from omigami.spectra_matching.gateways.fs_data_gateway import FSDataGateway, KEYS
+from omigami.spectra_matching.storage import FSDataGateway, KEYS
 
 ASSETS_DIR = Path(__file__).parents[0] / "assets"
 TEST_TASK_CONFIG = dict(max_retries=1, retry_delay=pd.Timedelta(seconds=0.1))
