+++ omigami/test/gateways/test_spectrum_gateway.py
@@ -5,7 +5,7 @@ from matchms.Spectrum import Spectrum
 from pytest_redis import factories
 
 from omigami.spec2vec.entities.embedding import Spec2VecEmbedding
-from omigami.spectra_matching.gateways import RedisSpectrumDataGateway
+from omigami.spectra_matching.storage import RedisSpectrumDataGateway
 from omigami.spec2vec.config import PROJECT_NAME
 from omigami.spec2vec.config import (
     SPECTRUM_ID_PRECURSOR_MZ_SORTED_SET,
