+++ omigami/test/ms2deepscore/flows/test_pretrained_flow.py
@@ -6,8 +6,7 @@ import pytest
 from drfs.filesystems import get_fs
 
 from omigami.config import MLFLOW_SERVER
-from omigami.spectra_matching.gateways.data_gateway import DataGateway
-from omigami.spectra_matching.gateways.fs_data_gateway import FSDataGateway
+from omigami.spectra_matching.storage import DataGateway, FSDataGateway
 from omigami.ms2deepscore.flows.pretrained_flow import (
     build_pretrained_flow,
     PretrainedFlowParameters,
