+++ omigami/test/ms2deepscore/tasks/test_process_spectrum_ms2deepscore.py
@@ -4,7 +4,7 @@ from unittest.mock import MagicMock
 import pytest
 from prefect import Flow
 
-from omigami.spectra_matching.gateways.fs_data_gateway import FSDataGateway
+from omigami.spectra_matching.storage import FSDataGateway
 from omigami.ms2deepscore.gateways.redis_spectrum_gateway import (
     MS2DeepScoreRedisSpectrumDataGateway,
 )
