+++ omigami/test/spec2vec/flows/test_training_flow.py
@@ -7,8 +7,7 @@ import pytest
 from drfs.filesystems import get_fs
 
 from omigami.config import SOURCE_URI_PARTIAL_GNPS
-from omigami.spectra_matching.gateways import RedisSpectrumDataGateway
-from omigami.spectra_matching.gateways.fs_data_gateway import FSDataGateway
+from omigami.spectra_matching.storage import RedisSpectrumDataGateway, FSDataGateway
 from omigami.spec2vec.config import PROJECT_NAME, SPEC2VEC_ROOT
 from omigami.spec2vec.flows.training_flow import (
     build_training_flow,
