+++ omigami/test/spec2vec/test_spec2vec_main.py
@@ -14,7 +14,7 @@ from omigami.config import (
 from omigami.deployer import FlowDeployer
 from omigami.spec2vec.factory import Spec2VecFlowFactory
 from omigami.spec2vec.main import run_spec2vec_flow
-from omigami.spectra_matching.gateways.fs_data_gateway import FSDataGateway
+from omigami.spectra_matching.storage import FSDataGateway
 from omigami.spectra_matching.tasks import DownloadData, DownloadParameters
 from omigami.test.conftest import monitor_flow_results
