+++ omigami/test/tasks/test_save_raw_spectra.py
@@ -4,8 +4,7 @@ import pytest
 from matchms.importing.load_from_json import as_spectrum
 from prefect import Flow
 
-from omigami.spectra_matching.gateways import RedisSpectrumDataGateway
-from omigami.spectra_matching.gateways.fs_data_gateway import FSDataGateway
+from omigami.spectra_matching.storage import RedisSpectrumDataGateway, FSDataGateway
 from omigami.spec2vec.config import SPECTRUM_ID_PRECURSOR_MZ_SORTED_SET
 from omigami.spectra_matching.spectrum_cleaner import SpectrumCleaner
 from omigami.spectra_matching.tasks import SaveRawSpectra, SaveRawSpectraParameters
